# public/openapi/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DebtResolve API",
        "contact": {
            "email": "jadiael@hotmail.com.br"
        },
        "license": {
            "name": "CC-BY-NC",
            "url": "https://creativecommons.org/licenses/by-nc/4.0/"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/charge-invitations/invitations": {
            "post": {
                "tags": [
                    "Charge Invitations"
                ],
                "summary": "Send charge invitation",
                "description": "Sends an invitation to participate in a charge",
                "operationId": "a65dff9f6cbd831e080d1dc8c1d7c127",
                "requestBody": {
                    "description": "Charge invitation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "charge_id"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email of the invitee",
                                        "type": "string"
                                    },
                                    "charge_id": {
                                        "description": "ID of the charge",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Invitation sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message about the operation status",
                                            "type": "string",
                                            "example": "Invitation to register and participate in billing sent successfully"
                                        },
                                        "data": {
                                            "description": "Data containing the charge and invitation details",
                                            "properties": {
                                                "charge": {
                                                    "$ref": "#/components/schemas/Charge"
                                                },
                                                "chargeInvitation": {
                                                    "$ref": "#/components/schemas/ChargeInvitation"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - User already participates in this charge",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message about the error",
                                            "type": "string",
                                            "example": "You already participate in this charge"
                                        },
                                        "errors": {
                                            "description": "Detailed error messages",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too Many Requests - Invitation limit for this email and billing exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message about the error",
                                            "type": "string",
                                            "example": "Invitation limit for this email and billing exceeded. Wait a week before sending a new invitation"
                                        },
                                        "errors": {
                                            "description": "Detailed error messages",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected error when creating resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message about the error",
                                            "type": "string",
                                            "example": "Unexpected error when creating resource"
                                        },
                                        "errors": {
                                            "description": "Detailed error messages",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/charge-invitations/process-charge-invitations/{token}": {
            "post": {
                "tags": [
                    "Charge Invitations"
                ],
                "summary": "Process Charge Invitations",
                "description": "Process the charge invitations based on the provided token.",
                "operationId": "3aeed32436a4a113814d0afd01038412",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Token associated with the charge invitation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful participation in the charge",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Congratulations! Now you participate in this charge"
                                        },
                                        "data": {
                                            "properties": {
                                                "charge": {
                                                    "$ref": "#/components/schemas/Charge"
                                                },
                                                "chargeInvitation": {
                                                    "$ref": "#/components/schemas/ChargeInvitation"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "Expired or invalid invitation link",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "This invite link has expired or does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Unable to participate in the charge",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "It is not possible to participate in this charge"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid request - trying to play both collector and debtor roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "You cannot play the role of collector and debtor at the same time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unexpected error occurred when processing the invitation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/charge-invitations/": {
            "get": {
                "tags": [
                    "Charge Invitations"
                ],
                "summary": "List Invitations",
                "description": "Retrieve a list of all invitations.",
                "operationId": "d287c38b4e33e953c660a66eb017368e",
                "responses": {
                    "200": {
                        "description": "Invitations listed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invitations listed successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "invitations": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Charge": {
                "properties": {
                    "id": {
                        "description": "Unique identifier for the charge",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the charge",
                        "type": "string",
                        "example": "Monthly rent"
                    },
                    "description": {
                        "description": "Description of the charge",
                        "type": "string",
                        "example": "Rent payment for the month"
                    },
                    "amount": {
                        "description": "Total amount of the charge",
                        "type": "number",
                        "example": 500
                    },
                    "installments_number": {
                        "description": "Number of installments for the charge",
                        "type": "integer",
                        "example": 12
                    },
                    "due_day": {
                        "description": "Due day for payment",
                        "type": "integer",
                        "example": 15
                    },
                    "collector_id": {
                        "description": "ID of the user who is the collector for this charge",
                        "type": "integer",
                        "example": 1
                    },
                    "debtor_id": {
                        "description": "ID of the user who is the debtor for this charge",
                        "type": "integer",
                        "example": 2
                    },
                    "collector": {
                        "$ref": "#/components/schemas/User"
                    },
                    "debtor": {
                        "$ref": "#/components/schemas/User"
                    },
                    "installments": {
                        "description": "List of installments associated with this charge",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Installment"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer"
            }
        }
    },
...
}